+++ CHANGELOG.md
@@ -11,13 +11,15 @@ All notable changes to TazUO will be recorded here.
 - Add sound API endpoints to LegionScripts - fpw
 - Added `API.ScriptName` and `API.ScriptPath`
 - Updated PSL browser UI and backend
-
+- Add .IsHidden to PyMobile in API
+- 
 ### Assistant
 - Added skills tab to Legion Assistant - Coryigon
 - Organizer tab now shows graphic when hovering over the graphic art
 - Add Mobile outline option - Highlighting mobiles by notoriety
 - Added TazUO chat(Top menu -> More -> TazUO Chat)
-
+- ItemDatabase search now defaults to not only "this character"
+- 
 ### Other
 - Move automatic py doc gen to tool usage
 - Added ibm-plex font to embedded fonts